三级分类菜单和品牌管理

+++ like-product/src/main/java/com/mdd/product/entity/PmsBrand.java
@@ -19,8 +19,8 @@ public class PmsBrand implements Serializable {
     private String name; // 品牌名
     private String logo; // 品牌logo地址
     private String descript; // 介绍
-    private Long showStatus; // 显示状态[0-不显示；1-显示]
+    private Integer showStatus; // 显示状态[0-不显示；1-显示]
     private String firstLetter; // 检索首字母
-    private Long sort; // 排序
+    private Integer sort; // 排序
 
 }
